Initial commit.
Moves all the existing code from the rails app to a new sinatra app.
Uses mongodb instead of postgres, since it looks like a better option
to the problem being resolved. The only class that really needed to be
changed was the saver (since its now called version_saver instead of
package_saver) and the version entity (package does not exist
anymore).